+++ HISTORY.md
@@ -1,5 +1,19 @@
 # History
 
+## 0.5.2
+
+### Changelog
+
+#### API
+
+- Fix Matplotlib `get_cmap` deprecation (#340)
+- Set `plot_centroids` to False when plotting (#339)
+- Raise error messages when `ask` is called without `ask_dqd` (#338)
+
+#### Documentation
+
+- Add BibTex citation for GECCO 2023 (#337)
+
 ## 0.5.1
 
 This release contains miscellaneous edits to our documentation from v0.5.0.